+++ server/utils/grpc_to_web.go
@@ -1,0 +1,49 @@
+package utils
+
+import (
+	"google.golang.org/grpc/codes"
+	"google.golang.org/grpc/status"
+	"net/http"
+	"strings"
+)
+
+func GrpcErrorToWebError(err error) ErrorResponse {
+	errorResp := ErrorResponse{Error: "Internal Server Error", StatusCode: http.StatusInternalServerError}
+
+	if strings.HasPrefix(err.Error(), "no device id") {
+		errorResp.Error = "Not Authorized"
+		errorResp.StatusCode = http.StatusForbidden
+	}
+
+	if s, ok := status.FromError(err); ok {
+		switch s.Code() {
+		case codes.NotFound:
+			errorResp.Error = "Not Found"
+			errorResp.StatusCode = http.StatusNotFound
+		case codes.Unimplemented:
+			errorResp.Error = "Not Implemented"
+			errorResp.StatusCode = http.StatusNotImplemented
+		case codes.InvalidArgument:
+			errorResp.Error = "Invalid Argument"
+			errorResp.StatusCode = http.StatusBadRequest
+		case codes.Internal:
+			errorResp.Error = "Internal Server Error"
+			errorResp.StatusCode = http.StatusInternalServerError
+		}
+
+		if s.Message() != "" {
+			errorResp.Error = s.Message()
+		}
+
+		if s.Details() != nil && len(s.Details()) > 0 {
+			errorResp.Details = s.Details()
+		}
+	}
+	return errorResp
+}
+
+type ErrorResponse struct {
+	Error      string        `json:"error"`
+	Details    []interface{} `json:"details,omitempty"`
+	StatusCode int           `json:"-"`
+}
